--- Skincare/Skincare/Views/Shelf/NewProductShelfViewController.swift
@@ -1,81 +1,0 @@
-//
-//  NewProductShelfViewController.swift
-//  Skincare
-//
-//  Created by Carolina Ortega on 14/12/21.
-//
-
-import Foundation
-import UIKit
-
-//DEVE RETORNAR TODA A API
-
-class NewProductShelfViewController: UIViewController{
-    @IBOutlet var searchBar: UISearchBar!
-    @IBOutlet var productTableView: UITableView!
-    var list: [String] = []
-    var searchProduct: [String]!
-    let defaults = UserDefaults.standard
-    var chosenProducts: [String] = []
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        navigationController?.setNavigationBarHidden(true, animated: false)
-        
-        //tableView
-        self.productTableView.delegate = self
-        self.productTableView.dataSource = self
-        
-        
-        //searchBar
-        searchBar.delegate = self
-        
-        
-    }
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        chosenProducts = defaults.stringArray(forKey: "myKey") ?? []
-        print(chosenProducts)
-        list = chosenProducts
-        navigationController?.setNavigationBarHidden(true, animated: animated)
-    }
-
-}
-
-extension NewProductShelfViewController: UISearchBarDelegate, UITableViewDataSource, UITableViewDelegate{
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return list.count
-    }
-    func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        return true
-    }
-    func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-    if editingStyle == .delete {
-//    Delete the row from the data source
-//    productTableView.deleteRows(at: [indexPath], with: .fade)
-    }
-//        else if editingStyle == .insert {
-//    Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-//    }
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "estante2", for: indexPath) as! ShelfTableViewCell
-        cell.textLabel?.text = list[indexPath.row]
-        return cell
-    }
-    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-        list = []
-        if searchText == ""{
-            list = chosenProducts
-        }else {
-            for product in chosenProducts{
-                if product.lowercased().contains(searchText.lowercased()){
-                    list.append(product)
-                }
-            }
-        }
-        self.productTableView.reloadData()
-    }
-}
